WIP:  Initial implementation of the burst mechanic for the data handlers.
The data handlers will now accumulate data in a burst sum,. and will
perfom actions on the burst sum at the end of the burst.  For now,
the corrector  and combiner classes will just calculate the average and
errors, and print them to screen.  *This will be very verbose!*

The correlator has empty fucntions for now; we may want to restructure how
the correlator actually stores its data to make it easier to have a burst
copy.

Eventaully we'd want these to connect to the burst trees, and not to be throwing
everything to the console.

This compiles, but I have NOT done any testing yet.